requirements.txt
Django==2.1.5
Pillow==5.4.1
pyocr==0.5.3
pytz==2018.9
six==1.12.0
